+++ src/chrono_vehicle/cosim/ChVehicleCosimBaseNode.cpp
@@ -395,29 +395,29 @@ void ChVehicleCosimBaseNode::SendGeometry(const ChVehicleGeometry& geom, int des
         const auto& normals = trimesh->GetCoordsNormals();
         const auto& idx_vertices = trimesh->GetIndicesVertexes();
         const auto& idx_normals = trimesh->GetIndicesNormals();
-        int nv = trimesh->GetNumVertices();
-        int nn = trimesh->GetNumNormals();
-        int nt = trimesh->GetNumTriangles();
+        unsigned int nv = trimesh->GetNumVertices();
+        unsigned int nn = trimesh->GetNumNormals();
+        unsigned int nt = trimesh->GetNumTriangles();
 
-        int surf_props[] = {nv, nn, nt, mesh.m_matID};
+        unsigned int surf_props[] = {nv, nn, nt, (unsigned int)mesh.m_matID};
         MPI_Send(surf_props, 4, MPI_INT, dest, 0, MPI_COMM_WORLD);
         if (m_verbose)
             cout << "[" << GetNodeTypeString() << "] Send: vertices = " << surf_props[0]
                  << "  triangles = " << surf_props[2] << endl;
 
         double* vert_data = new double[3 * nv + 3 * nn];
-        int* tri_data = new int[3 * nt + 3 * nt];
-        for (int iv = 0; iv < nv; iv++) {
+        unsigned int* tri_data = new unsigned int[3 * nt + 3 * nt];
+        for (unsigned int iv = 0; iv < nv; iv++) {
             vert_data[3 * iv + 0] = vertices[iv].x();
             vert_data[3 * iv + 1] = vertices[iv].y();
             vert_data[3 * iv + 2] = vertices[iv].z();
         }
-        for (int in = 0; in < nn; in++) {
+        for (unsigned int in = 0; in < nn; in++) {
             vert_data[3 * nv + 3 * in + 0] = normals[in].x();
             vert_data[3 * nv + 3 * in + 1] = normals[in].y();
             vert_data[3 * nv + 3 * in + 2] = normals[in].z();
         }
-        for (int it = 0; it < nt; it++) {
+        for (unsigned int it = 0; it < nt; it++) {
             tri_data[6 * it + 0] = idx_vertices[it].x();
             tri_data[6 * it + 1] = idx_vertices[it].y();
             tri_data[6 * it + 2] = idx_vertices[it].z();
